umbrella_movies/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.0
bcrypt==4.0.1
python-slugify==8.0.1